Chapter09/cloud_functions/readcitynames/requirements.txt
google-cloud-storage
functions-framework==3.*